Automatically ci-skip if all changed files are md
Currently, documentation changes require contributors to add a variation
of "ci skip" to their pull request in order to not run the full Rails
test suite. However, it requires contributors to know about this feature
in order to use it (to get faster/less flaky builds).

This commit tries to remove the requirements of that knowledge by
automatically skipping the full test suite in some cases (currently when
all changed files are markdown).

It also introduces a new opt-out "ci full" which can be used in cases
where the automatic opt-out is undesired for some reason.